Added support for particle radius

+++ verlet.cpp
@@ -22,24 +22,24 @@ namespace integrators
 	//Iterates across all spacial cordinates.
 	//Directly updates pts through the pointer.
 	//Use float return for debugging if needed.
-	float verlet::nextPosition(float index, float pos[], float vel[], mathTools::points* pts, physics::forces* f)
+	int verlet::nextPosition(float index, float pos[], float vel[], mathTools::points* pts, physics::forces* f)
 	{
 		float acc[3] = {0.0,0.0,0.0};
 		f->getAcceleration(pos,vel,systemTime,acc);
 		//Error code 0.0 -> No Error.
-		return 0.0;
+		return 0;
 	}
 
 	//Uses velocity verlet to get next velocity
 	//Iterates across all velocity cordinates.
 	//Directly updates pts through the pointer.
 	//Use float return for debugging if needed.
-	float verlet::nextVelocity(float index, float pos[], float vel[], mathTools::points* pts, physics::forces* f)
+	int verlet::nextVelocity(float index, float pos[], float vel[], mathTools::points* pts, physics::forces* f)
 	{
 		float acc[3] = {0.0,0.0,0.0};
 		f->getAcceleration(pos,vel,systemTime,acc);
 		//Error code 0.0 -> No Error.
-		return 0.0;
+		return 0;
 	}
 
 	//The velocity verlet Algorithms to obtain the next position term.
